Moves Media component to ui
